requirements.txt
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
google-genai>=1.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
html2text>=2024.2.26
